Macro-Shocks-Portfolio-Risk-Engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "macro-shock-risk-engine"
version = "1.0.0"
description = "Institutional-grade asymmetric macro shock risk detection and portfolio response engine"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Quantitative Research & Risk Engineering"},
]

dependencies = [
    "pydantic>=2.5",
    "numpy>=1.26",
    "pandas>=2.1",
    "structlog>=24.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]
nlp = [
    "sentence-transformers>=2.7",
    "torch>=2.1",
    "transformers>=4.36",
]
data = [
    "pandas-market-calendars>=4.4",
    "requests>=2.31",
    "aiohttp>=3.9",
    "fredapi>=0.5",
    "redis>=5.0",
]
backtest = [
    "scipy>=1.11",
    "statsmodels>=0.14",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "isort>=5.13",
    "mypy>=1.8",
    "ruff>=0.2",
    "ipykernel>=6.29",
    "jupyter>=1.0",
]
all = [
    "macro-shock-risk-engine[api,nlp,data,backtest,dev]",
]

[project.scripts]
msre = "macro_shock.__main__:main"

[tool.setuptools.packages.find]
where = ["python/src"]

[tool.setuptools.package-dir]
"" = "python/src"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["python/tests"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (require running services)",
    "slow: marks tests as slow-running",
]

[tool.coverage.run]
source = ["macro_shock"]
omit = ["*/tests/*", "*/notebooks/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
